weight_capacity is a float

+++ src/bionics.h
@@ -77,7 +77,7 @@ struct bionic_data {
     /**Factor modifiying weight capacity*/
     float weight_capacity_modifier;
     /**Bonus to weight capacity*/
-    int weight_capacity_bonus;
+    float weight_capacity_bonus;
     /**Amount of environemental protection offered by this bionic*/
     std::map<body_part, size_t> env_protec;
     /**
